# api/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"database/sql/driver"
	"fmt"

	"github.com/jackc/pgx/v5/pgtype"
)

type MatchType string

const (
	MatchTypeSINGLES MatchType = "SINGLES"
	MatchTypeDOUBLES MatchType = "DOUBLES"
	MatchTypeMIXED   MatchType = "MIXED"
)

func (e *MatchType) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = MatchType(s)
	case string:
		*e = MatchType(s)
	default:
		return fmt.Errorf("unsupported scan type for MatchType: %T", src)
	}
	return nil
}

type NullMatchType struct {
	MatchType MatchType `json:"match_type"`
	Valid     bool      `json:"valid"` // Valid is true if MatchType is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullMatchType) Scan(value interface{}) error {
	if value == nil {
		ns.MatchType, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.MatchType.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullMatchType) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.MatchType), nil
}

type EloHistory struct {
	ID                pgtype.UUID        `json:"id"`
	PlayerID          string             `json:"player_id"`
	EloBefore         int32              `json:"elo_before"`
	EloAfter          int32              `json:"elo_after"`
	MatchID           pgtype.UUID        `json:"match_id"`
	CreatedAt         pgtype.Timestamptz `json:"created_at"`
	PerformancePoints int32              `json:"performance_points"`
}

type Match struct {
	ID             pgtype.UUID        `json:"id"`
	MatchType      MatchType          `json:"match_type"`
	Team1Player1ID string             `json:"team1_player1_id"`
	Team1Player2ID pgtype.Text        `json:"team1_player2_id"`
	Team2Player1ID string             `json:"team2_player1_id"`
	Team2Player2ID pgtype.Text        `json:"team2_player2_id"`
	PlayedAt       pgtype.Timestamptz `json:"played_at"`
	Validated      bool               `json:"validated"`
	Set1Team1      int32              `json:"set1_team1"`
	Set1Team2      int32              `json:"set1_team2"`
	Set2Team1      int32              `json:"set2_team1"`
	Set2Team2      int32              `json:"set2_team2"`
	Set3Team1      pgtype.Int4        `json:"set3_team1"`
	Set3Team2      pgtype.Int4        `json:"set3_team2"`
}

type User struct {
	ID         string             `json:"id"`
	Email      string             `json:"email"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	Role       string             `json:"role"`
	FirstName  string             `json:"first_name"`
	LastName   string             `json:"last_name"`
	Gender     pgtype.Text        `json:"gender"`
	FfbadRank  pgtype.Text        `json:"ffbad_rank"`
	EloSingles int32              `json:"elo_singles"`
	EloDoubles int32              `json:"elo_doubles"`
	EloMixed   int32              `json:"elo_mixed"`
}