# api/docs/swagger.yaml
basePath: /
definitions:
  db.Match:
    properties:
      id:
        type: string
      match_type:
        $ref: '#/definitions/db.MatchType'
      played_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      set1_team1:
        type: integer
      set1_team2:
        type: integer
      set2_team1:
        type: integer
      set2_team2:
        type: integer
      set3_team1:
        $ref: '#/definitions/pgtype.Int4'
      set3_team2:
        $ref: '#/definitions/pgtype.Int4'
      team1_player1_id:
        type: string
      team1_player2_id:
        $ref: '#/definitions/pgtype.Text'
      team2_player1_id:
        type: string
      team2_player2_id:
        $ref: '#/definitions/pgtype.Text'
      validated:
        type: boolean
    type: object
  db.MatchType:
    enum:
    - SINGLES
    - DOUBLES
    - MIXED
    type: string
    x-enum-varnames:
    - MatchTypeSINGLES
    - MatchTypeDOUBLES
    - MatchTypeMIXED
  db.User:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      elo_doubles:
        type: integer
      elo_mixed:
        type: integer
      elo_singles:
        type: integer
      email:
        type: string
      ffbad_rank:
        $ref: '#/definitions/pgtype.Text'
      first_name:
        type: string
      gender:
        $ref: '#/definitions/pgtype.Text'
      id:
        type: string
      last_name:
        type: string
      role:
        type: string
    type: object
  handlers.createMatchRequest:
    properties:
      match_type:
        description: SINGLES | DOUBLES | MIXED
        type: string
      sets:
        description: 2 or 3 sets
        items:
          $ref: '#/definitions/handlers.setScore'
        type: array
      team1_player1_id:
        description: always required
        type: string
      team1_player2_id:
        description: required for DOUBLES / MIXED
        type: string
      team2_player1_id:
        description: always required
        type: string
      team2_player2_id:
        description: required for DOUBLES / MIXED
        type: string
    type: object
...
  handlers.setScore:
    properties:
      team1:
        type: integer
      team2:
        type: integer
    type: object
  handlers.syncRequest:
    properties:
      ffbad_rank:
        description: optional, NC | P12 | ... | N1
        type: string
      first_name:
        type: string
      gender:
        description: optional, MALE | FEMALE
        type: string
      last_name:
        type: string
    type: object
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Int4:
    properties:
      int32:
        format: int32
        type: integer
      valid:
        type: boolean
    type: object
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.Timestamptz:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for the BAP Pulse badminton club app — Bad A Paname.
  title: BAP Pulse API
  version: "1.0"
paths:
...
  /auth/sync:
    post:
      consumes:
      - application/json
      description: Called after a successful Firebase login. Creates the user profile
        on first login, or returns the existing one.
      parameters:
      - description: User first and last name (required on first login)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.syncRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.User'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Sync user profile
      tags:
      - auth
  /matches:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Match'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all matches
      tags:
      - matches
    post:
      consumes:
      - application/json
      description: Creates a match with set scores (best of 3, 21 points per set),
        recalculates ELO ratings, and records history.
      parameters:
      - description: Match details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createMatchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a match
      tags:
      - matches
  /matches/{id}:
    get:
      parameters:
      - description: Match ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Match'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a match
      tags:
      - matches
...
securityDefinitions:
  BearerAuth:
    description: 'Firebase ID Token — format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"